extract sphinxquery as lib
