+++ src/CMakeLists.txt
@@ -52,7 +52,7 @@ target_include_directories ( generated_config INTERFACE "${MANTICORE_BINARY_DIR}
 target_link_libraries ( lextra INTERFACE generated_config )
 
 # main bunch of sources (for libsphinx)
-add_library ( lmanticore STATIC sphinx.cpp sphinxquery.cpp sphinxutils.cpp
+add_library ( lmanticore STATIC sphinx.cpp sphinxutils.cpp
 		sphinxsort.cpp sortsetup.cpp sphinxexpr.cpp sphinxfilter.cpp sphinxsearch.cpp sphinxrt.cpp accumulator.cpp
 		sphinxjson.cpp sphinxaot.cpp sphinxplugin.cpp sphinxudf.c sphinxqcache.cpp 
 		attribute.cpp secondaryindex.cpp killlist.cpp searchnode.cpp json/cJSON.c sphinxpq.cpp
@@ -97,6 +97,9 @@ target_link_libraries ( lmanticore PUBLIC schema )
 
 add_subdirectory ( dict )
 target_link_libraries ( lmanticore PUBLIC dict )
+
+add_subdirectory ( sphinxquery )
+target_link_libraries ( lmanticore PUBLIC sphinxquery )
 
 if (WITH_ICU OR WITH_JIEBA)
 	add_library ( lcjk cjkpreprocessor.cpp )
@@ -163,7 +166,7 @@ target_link_libraries ( lmanticore PUBLIC lmanticore_coro )
 # So if you add headers to the project and NOT see them in visual studio solution - just list them here!
 set ( HEADERS sphinxfilter.h sphinxint.h sphinxpq.h sphinxrt.h
 		sphinxsort.h sphinxstem.h sphinxutils.h sphinxexpr.h sphinx.h sphinxjson.h sphinxplugin.h sphinxqcache.h
-		sphinxquery.h sphinxsearch.h sphinxstd.h sphinxudf.h lz4/lz4.h lz4/lz4hc.h http/http_parser.h secondaryindex.h
+		sphinxsearch.h sphinxstd.h sphinxudf.h lz4/lz4.h lz4/lz4hc.h http/http_parser.h secondaryindex.h
 		searchnode.h killlist.h attribute.h accumulator.h global_idf.h event.h threadutils.h threadutils_impl.h
 		hazard_pointer.h task_info.h mini_timer.h collation.h histogram.h sortsetup.h
 		indexsettings.h columnarlib.h fileio.h memio.h memio_impl.h queryprofile.h columnarfilter.h columnargrouper.h fileutils.h
@@ -185,32 +188,10 @@ set ( SEARCHD_H searchdaemon.h searchdconfig.h searchdddl.h searchdexpr.h search
 
 
 # add the extra targets in the case we want on-the-fly grammar compiler
-find_package ( BISON QUIET )
-set ( BISON_DIR "${MANTICORE_BINARY_DIR}/config" )
-set ( BIS_FLAGS "" )
-if (BISON_VERSION VERSION_GREATER 3.0)
-	set ( BIS_FLAGS "-Wno-deprecated" )
-endif ()
-
-function ( MY_BISON ParserName ParserSrc Dependency TargetBison )
-	if (NOT BISON_FOUND)
-		if (EXISTS "${BISON_DIR}/bis${ParserSrc}.c" AND EXISTS "${BISON_DIR}/bis${ParserSrc}.h")
-			infomsg ( "Will use pre-built ${BISON_DIR}/bis${ParserSrc}.c and ${BISON_DIR}/bis${ParserSrc}.h for grammar parser ${ParserName}" )
-			return ()
-		else ()
-			message ( FATAL_ERROR "No pre-compiled grammar files bis${ParserSrc}.c and bis${ParserSrc}.h for ${ParserSrc}.y exists, and Bison not found. Can't continue" )
-		endif ()
-	endif ()
-
-	LIST ( APPEND ${TargetBison}_BISON "${ParserSrc}.y" )
-	set ( ${TargetBison}_BISON ${${TargetBison}_BISON} PARENT_SCOPE )
-	BISON_TARGET ( ${ParserName} "${CMAKE_CURRENT_SOURCE_DIR}/${ParserSrc}.y" "${BISON_DIR}/bis${ParserSrc}.c" COMPILE_FLAGS ${BIS_FLAGS} )
-	set_source_files_properties ( ${Dependency} PROPERTIES OBJECT_DEPENDS ${BISON_${ParserName}_OUTPUT_SOURCE} SKIP_UNITY_BUILD_INCLUSION ON )
-endfunction ()
+include (mybison)
 
 MY_BISON ( ExprParser sphinxexpr sphinxexpr.cpp LMANTICORE )
 MY_BISON ( SelectParser sphinxselect sphinx.cpp LMANTICORE )
-MY_BISON ( QueryParser sphinxquery sphinxquery.cpp LMANTICORE )
 MY_BISON ( JsonParser sphinxjson sphinxjson.cpp LMANTICORE )
 MY_BISON ( SQLParser sphinxql searchdsql.cpp SEARCHD )
 MY_BISON ( DDLParser ddl searchdddl.cpp SEARCHD )
