+++ src/sphinxint.h
@@ -15,7 +15,7 @@
 
 #include "sphinxstd.h"
 #include "sphinxfilter.h"
-#include "sphinxquery.h"
+#include "sphinxquery/sphinxquery.h"
 #include "sphinxjsonquery.h"
 #include "sphinxutils.h"
 #include "fileio.h"
