+++ src/sphinxquery/sphinxquery.cpp
